noaa tests now in test_noaa_sample
